Create BeanCounter
This is code for the Bean Counter exercise found on page 60 in chapter 3. It asks for two different functions, one that determines the number of "B"s in a string, and another that lets a character argument determine the number of a certain character in a string. An interesting thing to note here is the case sensitivity of JavaScript.